Example datasets {pbdPROF}	R Documentation

Small example datasets

Description

Sample profiler outputs for testing and package demonstration.

Format

Each dataset contains information profiled by fpmpi for pbdMPI/demo/allreduce.fpmpi.r, pbdDMAT/demo/svd.fpmpi.r, and Rmpi/demo/masterslavePI.fpmpi.r.

Each dataset contains information profiled by mpiP for pbdMPI/demo/allreduce.mpip.r, pbdDMAT/demo/svd.mpip.r, and Rmpi/demo/masterslavePI.mpip.r.

Details

Several sample profiler outputs are provided with the package, and utilized in the package demos to show off the functionality of the parsing and plotting methods provided by the package. These small datasets are located in the pbdPROF/extdata/ folder of the installed package (and pbdPROF/inst/extdata/ in the package source).

The example datasets are profiled in 2 processors.

[Package pbdPROF version 0.3-0 Index]