pbdPROF-package {pbdPROF}	R Documentation

MPI Profiling Tools

Description

This package contains several libraries for profiling MPI codes, as well as some R tools for parsing, analyzing, and plotting their outputs.

Details

Package: 	 pbdPROF

Type: 	 Package

License: 	 MPL

LazyLoad: 	 yes

This package requires an MPI library (OpenMPI, MPICH2, or LAM/MPI).

Author(s)

Wei-Chen Chen, Drew Schmidt, Gaurav Sehrawat, Pragneshkumar Patel, and George Ostrouchov.

References

Programming with Big Data in R Website: http://r-pbd.org/

Examples

## Not run: 
demo(allreduce.fpmpi, "pbdPROF")
demo(svd.fpmpi, "pbdPROF")
demo(masterslavePI.fpmpi, "pbdPROF")
demo(allreduce.mpip, "pbdPROF")
demo(svd.mpip, "pbdPROF")
demo(masterslavePI.mpip, "pbdPROF")

## End(Not run)

[Package pbdPROF version 0.3-0 Index]