prof-class {pbdPROF}	R Documentation

Class prof

Description

Class for Profiler Output

Slots

profiler

The type of profiler used (e.g. fpmpi, mpiP). Stored as a character vector.

raw

The raw (non-processed) profiler output. Storage is basically a character vector, but set as a virtual class rawprof.

parsed

A dataframe containing the processed version of the raw data.

See Also

read.prof

[Package pbdPROF version 0.3-0 Index]