profilers {pbdPROF}	R Documentation

Profilers

Description

Profilers Directly Supported by pbdPROF

Details

Currently, the fpmpi library is fully supported; a version of this library is bundled with the source of the pbdPROF package (see package vignette for more details). Additional libraries can easily be linked with pbdPROF, but these are not yet fully supported. The mpiP and TAU profilers are expected to be fully supported by the conclusion of Google Summer of Code (~September 30 2013).

References

Programming with Big Data in R Website: http://r-pbd.org/

fpmpi website: http://www.mcs.anl.gov/research/projects/fpmpi/WWW/

mpiP website: http://mpip.sourceforge.net/

TAU website: http://www.cs.uoregon.edu/research/tau/home.php

[Package pbdPROF version 0.3-0 Index]