read.prof {pbdPROF}	R Documentation

Reading Profiling Outputs

Description

Reader for profiler outputs.

Usage

read.prof(file.name, ...)

Arguments

file.name
a full file name.

...
options for readLines

Details

This function reads in profiling outputs from MPI-using R code and stores the output in a prof class object. The reading is managed by the base::readLines() function. The user does not need to specify the type of profiler output being used (e.g., whether the profiler text is from fpmpi, mpiP, etc.).

Additionally, this method automatically parses the output into a condensed, manageable dataframe (the parsed slot of the prof class).

Value

A prof class object.

See Also

prof-class

Examples

## Not run: 
library(pbdPROF)

fn <- system.file("data/allreduce.fpmpi", package = "pbdPROF")
da <- read.prof(fn, lib.type = "fpmpi")

da

## End(Not run)

[Package pbdPROF version 0.3-0 Index]